repurpose Main component to simplify boot.jsx
